requirements.txt
pandas
python-dotenv
numpy
pymongo[srv]==3.12
pymongo
certifi

# -e . #this is only used when u want to use pacakages as deploy on pypy or cloud